requirements.txt
pytorch
opencv
torchvision
onnx
onnxruntime
onnx-simplifier
pillow